Rename 알고리즘 스터디 - 1일 될 때까지(그리디).cpp to 1일 될 때까지(그리디).cpp
